implented pid loop

+++ software/src/include/mcp9600.h
@@ -109,9 +109,9 @@
 
 uint8_t mcp9600_init();
 
-double mcp9600_get_hj_temp();
-double mcp9600_get_cj_temp();
-double mcp9600_get_temp_delta();
+float mcp9600_get_hj_temp();
+float mcp9600_get_cj_temp();
+float mcp9600_get_temp_delta();
 
 uint32_t mcp9600_get_adc();
 
@@ -132,6 +132,6 @@ uint8_t mcp9600_get_alert_config(uint8_t ubAlert);
 
 void mcp9600_set_alert_hysteresis(uint8_t ubAlert, uint8_t ubHysteresis);
 
-void mcp9600_set_alert_limit(uint8_t ubAlert, double dLimit);
+void mcp9600_set_alert_limit(uint8_t ubAlert, float dLimit);
 
 #endif  // __MCP9600_H__
